feat(wabe): allow to mutate an object without return any data (#121)
We need to be able to create an object but don't return any data, like id. Just add a "ok" field that equal to true if all is good. It avoid to get useless data you don't need for just to match with GraphQL requirement. It also allows to avoid issues when for example anonymous have the permission to create an object but not to read.